bug fix: querying instances that are not running and empty instance_id.txt

+++ utils.c
@@ -1,5 +1,6 @@
 #include <stdio.h>
 #include <stdlib.h>
+#include <string.h>
 
 // Warning: if you change the value of MAX_ID_LEN, you might need to change the
 // length of the format_specifier string.
@@ -7,7 +8,7 @@
 #define MAX_ID_LEN 32   // The maximum length of an instance ID
 #define MAX_NAME_LEN 64 // The maximum length of an instance name
 
-/*
+/**
  * Terminates an AWS instance.
  *
  * @param instance_name The name of the instance to terminate. Max length is 64.
@@ -20,38 +21,38 @@ void terminate_aws_instance(const char *instance_name)
     char instance_id[MAX_ID_LEN + 1];
 
     // Length of the AWS command plus the maximum length of the instance name
-    char get_id_command[127 + (MAX_NAME_LEN + 1)];
+    char get_id_command[187 + (MAX_NAME_LEN + 1)];
 
-    // Length of the file write command, including the length of the AWS command 
-    // and instance name
-    char to_file[19 + (127 + MAX_NAME_LEN + 1)];
-
-    // Length of the command to terminate the instance plus the max length of 
+    // Length of the command to terminate the instance plus the max length of
     // the instance ID
     char command[58 + (MAX_ID_LEN + 1)];
 
-    // (% + s + null terminator) + length of the instance ID string 
+    // (% + s + null terminator) + character length of MAX_ID_LEN
     char format_specifier[3 + 2];
 
     snprintf(format_specifier, sizeof(format_specifier), "%%%ds", MAX_ID_LEN); // "%ds"
 
-    snprintf(get_id_command, sizeof(get_id_command), "aws ec2 describe-instances --filters \"Name=tag:Name,Values=%s\" --query \"Reservations[*].Instances[*].InstanceId\" --output text", instance_name);
+    snprintf(get_id_command, sizeof(get_id_command), "aws ec2 describe-instances --filters \"Name=tag:Name,Values=%s\" \"Name=instance-state-name,Values=running\" --query \"Reservations[*].Instances[*].InstanceId\" --output text > instance_id.txt", instance_name);
 
-    snprintf(to_file, sizeof(to_file), "%s%s", get_id_command, " > instance_id.txt");
-
-    system(to_file);
+    system(get_id_command);
 
     fd = fopen("instance_id.txt", "r");
-    if (NULL != fd)
+    if (NULL != fd) // If the file exists
     {
-        fscanf(fd, format_specifier, instance_id);
+        fseek(fd, 0, SEEK_END);
+        if (ftell(fd) != 0) // If the file is not empty
+        {
+            fseek(fd, 0, SEEK_SET);
 
-        fclose(fd);
+            fscanf(fd, format_specifier, instance_id);
 
-        snprintf(command, sizeof(command), "aws ec2 terminate-instances --instance-ids %s > /dev/null", instance_id);
+            fclose(fd);
 
-        system(command);
+            snprintf(command, sizeof(command), "aws ec2 terminate-instances --instance-ids %s > /dev/null", instance_id);
 
-        remove("instance_id.txt");
+            system(command);
+
+            remove("instance_id.txt");
+        }
     }
 }
